DM-8385: PDAC v1 deployment  - setup LSST/suit repo  - added lsst logo onto the banner  - make sure build scripts are fully functional  - minor code refactoring
