+++ package.json
@@ -52,7 +52,6 @@
     "style-loader": "^0.12.3",
     "url-loader": "^0.5.7",
     "file-loader": "^0.9.0",
-    "html-webpack-plugin": "^1.6.1",
     "extract-text-webpack-plugin": "^0.8.0",
     "eslint"            : "^3.4.0",
     "eslint-loader"     : "^1.5.0",
