+++ src/firefly/js/ui/Banner.css
@@ -9,8 +9,6 @@
 
 .banner__left {
     flex-grow: 0;
-    width: 75px;
-    height: 75px;
 }
 
 .banner__middle {
